requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.9
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.4
